+++ lib/DocumentService.php
@@ -137,6 +137,7 @@ class DocumentService {
         }
 
         $document_revision_id = self::generateRevisionId($document_revision_id);
+        $urlToConverter = $urlToConverter . "?shardKey=" . $document_revision_id;
 
         if (empty($from_extension)) {
             $from_extension = pathinfo($document_uri)["extension"];